Rework autosave

+++ src/gui/src/MainForm.cpp
@@ -84,8 +84,9 @@ using namespace H2Core;
 
 int MainForm::sigusr1Fd[2];
 
-MainForm::MainForm( QApplication * pQApplication )
+MainForm::MainForm( QApplication * pQApplication, QString sSongFilename )
 	: QMainWindow( nullptr )
+	, m_sPreviousAutoSaveFilename( "" )
 {
 	auto pPref = H2Core::Preferences::get_instance();
 	
@@ -103,6 +104,30 @@ MainForm::MainForm( QApplication * pQApplication )
 	m_pQApp = pQApplication;
 
 	m_pQApp->processEvents();
+
+	// When using the Non Session Management system, the new Song
+	// will be loaded by the NSM client singleton itself and not
+	// by the MainForm. The latter will just access the already
+	// loaded Song.
+	if ( ! H2Core::Hydrogen::get_instance()->isUnderSessionManagement() ){
+		std::shared_ptr<H2Core::Song>pSong = nullptr;
+
+		if ( sSongFilename.isEmpty() ) {
+			if ( pPref->isRestoreLastSongEnabled() ) {
+				sSongFilename = pPref->getLastSongFilename();
+			}
+		}
+
+		bool bRet;
+		if ( !sSongFilename.isEmpty() ) {
+			bRet = HydrogenApp::openSong( sSongFilename );
+		}
+
+		if ( sSongFilename.isEmpty() || ! bRet ) {
+			pSong = H2Core::Song::getEmptySong();
+			HydrogenApp::openSong( pSong );
+		}
+	}
 
 	QFont font( pPref->getApplicationFontFamily(), getPointSize( pPref->getFontSize() ) );
 	setFont( font );
@@ -133,7 +158,7 @@ MainForm::MainForm( QApplication * pQApplication )
 	installEventFilter( this );
 
 	connect( &m_AutosaveTimer, SIGNAL(timeout()), this, SLOT(onAutoSaveTimer()));
-	m_AutosaveTimer.start( 60 * 1000 );
+	startAutosaveTimer();
 
 #ifdef H2CORE_HAVE_LASH
 
@@ -183,9 +208,14 @@ MainForm::MainForm( QApplication * pQApplication )
 
 MainForm::~MainForm()
 {
-	// remove the autosave file
-	QFile file( getAutoSaveFilename() );
-	file.remove();
+	auto pHydrogen = Hydrogen::get_instance();
+	
+	// Remove the autosave file in case all modifications already have
+	// been written to disk.
+	if ( ! pHydrogen->getSong()->getIsModified() ) {
+		QFile file( getAutoSaveFilename() );
+		file.remove();
+	}
 
 	//if a playlist is used, we save the last playlist-path to hydrogen.conf
 	Preferences::get_instance()->setLastPlaylistFilename( Playlist::get_instance()->getFilename() );
@@ -193,13 +223,7 @@ MainForm::~MainForm()
 	if ( (Hydrogen::get_instance()->getAudioEngine()->getState() == H2Core::AudioEngine::State::Playing) ) {
 		Hydrogen::get_instance()->sequencer_stop();
 	}
-
-	// remove the autosave file
-	m_AutosaveTimer.stop();
-	QFile autosaveFile( "hydrogen_autosave.h2song" );
-	autosaveFile.remove();
-
-
+	
 	hide();
 
 	if (h2app != nullptr) {
@@ -424,6 +448,22 @@ void MainForm::createMenuBar()
 	//~ INFO menu
 }
 
+void MainForm::startAutosaveTimer() {
+	int nAutosavesPerHour = Preferences::get_instance()->m_nAutosavesPerHour;
+
+	if ( nAutosavesPerHour > 0 ) {
+		if ( nAutosavesPerHour > 360 ) {
+			ERRORLOG( QString( "Too many autosaves per hour set [%1]. Using 360 - once a second - instead." )
+					  .arg( nAutosavesPerHour ) );
+			nAutosavesPerHour = 360;
+		}
+		m_AutosaveTimer.start( std::round( 60 * 60 * 1000 /
+										   static_cast<float>(nAutosavesPerHour) ) );
+	} else {
+		DEBUGLOG( "Autosave disabled" );
+	}
+}
+
 void MainForm::onLashPollTimer()
 {
 #ifdef H2CORE_HAVE_LASH
@@ -546,7 +586,6 @@ void MainForm::action_file_new()
 		return;
 	}
 	
-	h2app->m_pUndoStack->clear();
 	pHydrogen->getTimeline()->deleteAllTempoMarkers();
 	pHydrogen->getTimeline()->deleteAllTags();
 	std::shared_ptr<Song> pSong = Song::getEmptySong();
@@ -620,7 +659,7 @@ void MainForm::action_file_save_as()
 	}
 	else {
 		QFileInfo fileInfo( lastFilename );
-		defaultFilename = fileInfo.baseName();
+		defaultFilename = fileInfo.completeBaseName();
 	}
 	defaultFilename += Filesystem::songs_ext;
 
@@ -829,43 +868,19 @@ void MainForm::action_file_export_pattern_as()
 }
 
 void MainForm::action_file_open() {
-		
-	if ( Hydrogen::get_instance()->getAudioEngine()->getState() == H2Core::AudioEngine::State::Playing ) {
-		Hydrogen::get_instance()->sequencer_stop();
-	}
-
-	bool bProceed = handleUnsavedChanges();
-	if( !bProceed ) {
-		return;
-	}
-
 	QString sPath = Preferences::get_instance()->getLastOpenSongDirectory();
 	if ( ! Filesystem::dir_readable( sPath, false ) ){
 		sPath = Filesystem::songs_dir();
 	}
 
-	QFileDialog fd(this);
-	fd.setFileMode( QFileDialog::ExistingFile );
-	fd.setDirectory( sPath );
-	fd.setNameFilter( Filesystem::songs_filter_name );
-
+	QString sWindowTitle;
 	if ( H2Core::Hydrogen::get_instance()->isUnderSessionManagement() ) {
-		fd.setWindowTitle( tr( "Import song into Session" ) );
+		sWindowTitle = tr( "Import song into Session" );
 	} else {
-		fd.setWindowTitle( tr( "Open song" ) );
+		sWindowTitle = tr( "Open song" );
 	}
 
-	QString sFilename;
-	if ( fd.exec() == QDialog::Accepted ) {
-		Preferences::get_instance()->setLastOpenSongDirectory( fd.directory().absolutePath() );
-		sFilename = fd.selectedFiles().first();
-	}
-
-	if ( !sFilename.isEmpty() ) {
-		HydrogenApp::get_instance()->openSong( sFilename );
-	}
-
-	HydrogenApp::get_instance()->getInstrumentRack()->getSoundLibraryPanel()->update_background_color();
+	openSongWithDialog( sWindowTitle, sPath, false );
 }
 
 
@@ -903,45 +918,53 @@ void MainForm::action_file_openPattern()
 	}
 }
 
-/// \todo parametrizzare il metodo action_file_open ed eliminare il seguente...
-void MainForm::action_file_openDemo()
-{
-	if ( Hydrogen::get_instance()->getAudioEngine()->getState() == H2Core::AudioEngine::State::Playing ) {
-		Hydrogen::get_instance()->sequencer_stop();
+void MainForm::action_file_openDemo() {
+	QString sWindowTitle;
+	if ( ! H2Core::Hydrogen::get_instance()->isUnderSessionManagement() ) {
+		sWindowTitle = tr( "Open Demo Song" );
+	} else {
+		sWindowTitle = tr( "Import Demo Song into Session" );
 	}
 
-	bool proceed = handleUnsavedChanges();
-	if(!proceed) {
+	openSongWithDialog( sWindowTitle, Filesystem::demos_dir(), true );
+}
+
+void MainForm::openSongWithDialog( const QString& sWindowTitle, const QString& sPath, bool bIsDemo ) {
+	auto pHydrogen = Hydrogen::get_instance();
+	if ( pHydrogen->getAudioEngine()->getState() ==
+		 H2Core::AudioEngine::State::Playing ) {
+		pHydrogen->sequencer_stop();
+	}
+
+	bool bProceed = handleUnsavedChanges();
+	if( !bProceed ) {
 		return;
 	}
 
-	h2app->m_pUndoStack->clear();
 	QFileDialog fd(this);
-	fd.setFileMode(QFileDialog::ExistingFile);
+	fd.setFileMode( QFileDialog::ExistingFile );
+	fd.setDirectory( sPath );
 	fd.setNameFilter( Filesystem::songs_filter_name );
+	fd.setWindowTitle( sWindowTitle );
 
-	if ( ! H2Core::Hydrogen::get_instance()->isUnderSessionManagement() ) {
-		fd.setWindowTitle( tr( "Open Demo Song" ) );
-	} else {
-		fd.setWindowTitle( tr( "Import Demo Song into Session" ) );
+	QString sFilename;
+	if ( fd.exec() == QDialog::Accepted ) {
+		if ( ! bIsDemo ) {
+			Preferences::get_instance()->setLastOpenSongDirectory( fd.directory().absolutePath() );
+		}
+		sFilename = fd.selectedFiles().first();
 	}
 
-	fd.setDirectory( Filesystem::demos_dir() );
-
-	QString filename;
-	if (fd.exec() == QDialog::Accepted) {
-		filename = fd.selectedFiles().first();
-	}
-
-	if ( !filename.isEmpty() ) {
-		HydrogenApp::get_instance()->openSong( filename );
-		if ( ! H2Core::Hydrogen::get_instance()->isUnderSessionManagement() ) {
-			Hydrogen::get_instance()->getSong()->setFilename( "" );
+	if ( !sFilename.isEmpty() ) {
+		HydrogenApp::get_instance()->openSong( sFilename );
+		if ( bIsDemo &&
+			 ! pHydrogen->isUnderSessionManagement() ) {
+			pHydrogen->getSong()->setFilename( "" );
 		}
 	}
+
+	HydrogenApp::get_instance()->getInstrumentRack()->getSoundLibraryPanel()->update_background_color();
 }
-
-
 
 void MainForm::showPreferencesDialog()
 {
@@ -1392,6 +1415,10 @@ void MainForm::onPreferencesChanged( H2Core::Preferences::Changes changes ) {
 
 	if ( changes & H2Core::Preferences::Changes::Colors ) {
 		Skin::setPalette( m_pQApp );
+	}
+
+	if ( changes & H2Core::Preferences::Changes::GeneralTab ) {
+		startAutosaveTimer();
 	}
 }
 	
@@ -2001,26 +2028,62 @@ QString MainForm::getAutoSaveFilename()
 	std::shared_ptr<Song> pSong = Hydrogen::get_instance()->getSong();
 	assert( pSong );
 	QString sOldFilename = pSong->getFilename();
-	QString newName = "autosave.h2song";
+	QString sNewName;
 
 	if ( !sOldFilename.isEmpty() ) {
-		newName = sOldFilename.left( sOldFilename.length() - 7 ) + ".autosave.h2song";
+
+		QFileInfo fileInfo( sOldFilename );
+
+		// In case the user did open a hidden file, the baseName()
+		// will be an empty string.
+		QString sBaseName( fileInfo.completeBaseName() );
+		if ( sBaseName.front() == "." ) {
+			sBaseName.remove( 0, 1 );
+		}
+
+		QString sAbsoluteDir( fileInfo.absoluteDir().absolutePath() );
+		if ( ! Filesystem::file_writable( sOldFilename, true ) ) {
+
+			sNewName = QString( "%1%2.autosave.h2song" )
+				.arg( Filesystem::songs_dir() ).arg( sBaseName );
+		
+			WARNINGLOG( QString( "Path of current song [%1] is not writable. Autosave will store the song as [%2] instead." )
+						.arg( sOldFilename ).arg( sNewName ) );
+		} else {
+			sNewName = QString( "%1/.%2.autosave.h2song" )
+				.arg( sAbsoluteDir ).arg( sBaseName );
+		}
+	} else {
+		// Store the default autosave file in the user's song data
+		// folder to not clutter their working directory.
+		sNewName = QString( "%1autosave.h2song" )
+			.arg( Filesystem::songs_dir() );
 	}
 
-	return newName;
+	return sNewName;
 }
 
 
 
 void MainForm::onAutoSaveTimer()
 {
-	//INFOLOG( "[onAutoSaveTimer]" );
 	auto pHydrogen = Hydrogen::get_instance();
 	std::shared_ptr<Song> pSong = pHydrogen->getSong();
+
 	assert( pSong );
 	if ( pSong->getIsModified() ) {
 		QString sOldFilename = pSong->getFilename();
-		pSong->save( getAutoSaveFilename() );
+
+		QString sAutoSaveFilename = getAutoSaveFilename();
+		if ( sAutoSaveFilename != m_sPreviousAutoSaveFilename ) {
+			if ( ! m_sPreviousAutoSaveFilename.isEmpty() ) {
+				QFile file( m_sPreviousAutoSaveFilename );
+				file.remove();
+			}
+			m_sPreviousAutoSaveFilename = sAutoSaveFilename;
+		}
+			
+		pSong->save( sAutoSaveFilename );
 
 		pSong->setFilename( sOldFilename );
 		pHydrogen->setIsModified( true );
@@ -2250,6 +2313,13 @@ void MainForm::action_banks_properties()
 	// Cleaning up the last pInfo we did not deleted due to the break
 	// statement.
 	delete pDrumkitInfo;
+}
+
+void MainForm::updateSongEvent( int nValue ) {
+	if ( nValue == 0 ) {
+		// A new song was set.
+		updateRecentUsedSongList();
+	}
 }
 
 void MainForm::startPlaybackAtCursor( QObject* pObject ) {
